Base58*Test: move to `base` package
Base58 is already there.
