+++ core/src/test/java/org/bitcoinj/base/Base58EncodeCheckedTest.java
@@ -15,9 +15,9 @@
  * limitations under the License.
  */
 
-package org.bitcoinj.core;
+package org.bitcoinj.base;
 
-import org.bitcoinj.base.Base58;
+import org.bitcoinj.core.LegacyAddress;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
